Fixing the build and refactoring it
